pyproject.toml
[tool.poetry]
name = "todo"
version = "0.1.0"
description = "This is the TODO CRUD application for Codavatar"
authors = ["Aayush Dip Giri <aayush.giri@codavatar.tech>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.100.0"
sqlalchemy = "^2.0.19"
uvicorn = "^0.23.1"
psycopg2-binary = "^2.9.6"
python-dotenv = "^1.0.0"
bcrypt = "^4.0.1"
annotated = "^0.0.2"
alembic = "^1.11.1"
fastapi-sqlalchemy = "^0.2.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
pydantic-settings = "^2.0.2"
python-multipart = "^0.0.6"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
